change lat lon to floats in database, drop database

+++ db/migrate/20150220004208_add_lat_lon_to_venues.rb
@@ -1,6 +1,6 @@
 class AddLatLonToVenues < ActiveRecord::Migration
   def change
-    add_column :venues, :lat, :integer
-    add_column :venues, :lon, :integer
+    add_column :venues, :lat, :float
+    add_column :venues, :lon, :float
   end
 end
